[reviewed] Modified README and C files * rcv -> recv variable issue fixed * ifs with single line were unwrapped from braces * comments are now only C-style (/* */) * changed DEFAULT_LOCAL_FILE to EXAMPLE_LOCAL_FILE * if (sender) instead of if (sender == 1) * if !strcmp() instead of strcmp() == 0 * renamed files to make them distinguishable from sample apps * added response examples for usage * modified hugepages recommendations
