+++ tests/validation/recver_val.c
@@ -4,10 +4,8 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
-#include <assert.h>
 #include <bsd/string.h>
 #include <getopt.h>
-#include <linux/limits.h>
 #include <signal.h>
 #include <stdbool.h>
 #include <stdio.h>
@@ -19,39 +17,10 @@
 #include "common.c"
 
 static volatile bool keepRunning = true;
-bool loop = DEFAULT_INFINITE_LOOP;
 
 void intHandler(int dummy)
 {
     keepRunning = 0;
-}
-
-int read_test_data(FILE* fp, MeshBuffer* buf, uint32_t frame_size)
-{
-    int ret = 0;
-
-    assert(fp != NULL && buf != NULL);
-    assert(buf->data_len >= frame_size);
-
-    if (fread(buf->data, frame_size, 1, fp) < 1) {
-        ret = -1;
-    }
-    return ret;
-}
-
-int gen_test_data(MeshBuffer *buf, uint32_t frame_count)
-{
-    /* operate on the buffer */
-    void* ptr = buf->data;
-
-    /* frame counter */
-    *(uint32_t*)ptr = frame_count;
-    ptr += sizeof(frame_count);
-
-    /* timestamp */
-    clock_gettime( CLOCK_REALTIME , (struct timespec*)ptr);
-
-    return 0;
 }
 
 int main(int argc, char** argv)
@@ -63,7 +32,7 @@ int main(int argc, char** argv)
     char protocol_type[32] = DEFAULT_PROTOCOL;
     char payload_type[32] = DEFAULT_PAYLOAD_TYPE;
 
-    static char file_name[128] = "";
+    char file_name[128] = "";
 
     char pix_fmt_string[32] = DEFAULT_VIDEO_FMT;
     char socket_path[108] = DEFAULT_MEMIF_SOCKET_PATH;
@@ -73,7 +42,6 @@ int main(int argc, char** argv)
     uint32_t width = DEFAULT_FRAME_WIDTH;
     uint32_t height = DEFAULT_FRAME_HEIGHT;
     double vid_fps = DEFAULT_FPS;
-    uint32_t total_num = 300;
 
     MeshClient *client;
     MeshConnection *conn;
@@ -90,9 +58,9 @@ int main(int argc, char** argv)
         { "height", required_argument, NULL, 'h' },
         { "fps", required_argument, NULL, 'f' },
         { "pix_fmt", required_argument, NULL, 'x' },
-        
-        { "rcv_ip", required_argument, NULL, 'r' },
-        { "rcv_port", required_argument, NULL, 'i' },
+
+        { "recv_ip", required_argument, NULL, 'r' },
+        { "recv_port", required_argument, NULL, 'i' },
         { "send_ip", required_argument, NULL, 's' },
         { "send_port", required_argument, NULL, 'p' },
         
@@ -100,17 +68,15 @@ int main(int argc, char** argv)
         { "payload_type", required_argument, NULL, 't' },
         { "socketpath", required_argument, NULL, 'k' },
         { "interfaceid", required_argument, NULL, 'd' },
-
-        { "loop", required_argument, NULL, 'l' },
+        
         { 0 }
     };
 
     /* infinite loop, to be broken when we are done parsing options */
     while (1) {
-        opt = getopt_long(argc, argv, "Hb:w:h:f:x:r:i:s:p:o:t:k:d:l:", longopts, 0);
-        if (opt == -1) {
+        opt = getopt_long(argc, argv, "Hb:w:h:f:x:r:i:s:p:o:t:k:d:", longopts, 0);
+        if (opt == -1)
             break;
-        }
 
         switch (opt) {
         case 'H':
@@ -154,9 +120,6 @@ int main(int argc, char** argv)
             break;
         case 'd':
             interface_id = atoi(optarg);
-            break;
-        case 'l':
-            loop = (atoi(optarg)>0);
             break;
         default:
             break;
@@ -272,7 +235,7 @@ int main(int argc, char** argv)
         goto error_delete_conn;
     }
 
-    err = mesh_establish_connection(conn, MESH_CONN_KIND_SENDER);
+    err = mesh_establish_connection(conn, MESH_CONN_KIND_RECEIVER);
     if (err) {
         printf("Failed to establish connection: %s (%d)\n",
                mesh_err2str(err), err);
@@ -284,71 +247,68 @@ int main(int argc, char** argv)
     uint32_t frame_count = 0;
     uint32_t frame_size = conn->buf_size;
 
-    FILE* input_fp = NULL;
+    FILE* dump_fp = NULL;
 
     const uint32_t stat_interval = 10;
     double fps = 0.0;
     double throughput_MB = 0;
     double stat_period_s = 0;
 
+    struct timespec ts_recv = {}, ts_send = {};
     struct timespec ts_begin = {}, ts_end = {};
-    struct timespec ts_frame_begin = {}, ts_frame_end = {};
 
-    if (strlen(file_name) > 0) {
-        struct stat statbuf = { 0 };
-        if (stat(file_name, &statbuf) == -1) {
-            perror(NULL);
-            goto error_delete_conn;
-        }
+    void *ptr = NULL;
+    int timeout = MESH_TIMEOUT_INFINITE;
+    bool first_frame = true;
+    float latency = 0;
 
-        input_fp = fopen(file_name, "rb");
-        if (input_fp == NULL) {
-            printf("Failed to open input file: %s\n", file_name);
-            goto error_delete_conn;
-        }
-    }
+    if (strlen(file_name) > 0)
+        dump_fp = fopen(file_name, "wb");
 
-    const __useconds_t pacing = 1000000.0 / vid_fps;
     while (keepRunning) {
-        /* Timestamp for frame start. */
-        clock_gettime(CLOCK_REALTIME, &ts_frame_begin);
+        /* receive frame */
+        if (first_frame) {
+            /* infinite for the 1st frame. */
+            timeout = MESH_TIMEOUT_INFINITE;
+        } else {
+            /* 1 second */
+            timeout = 1000;
+        }
 
-        err = mesh_get_buffer(conn, &buf);
+        err = mesh_get_buffer_timeout(conn, &buf, timeout);
+        if (err == -MESH_ERR_CONN_CLOSED) {
+            printf("Connection closed\n");
+            break;
+        }
         if (err) {
             printf("Failed to get buffer: %s (%d)\n", mesh_err2str(err), err);
             break;
         }
 
-        printf("INFO: frame_size = %u\n", frame_size);
+        printf("INFO: buf->len = %ld frame size = %u\n", buf->data_len, frame_size);
 
-        if (input_fp == NULL) {
-            gen_test_data(buf, frame_count);
-        } else {
-            if (read_test_data(input_fp, buf, frame_size) < 0) {
-                if (input_fp != NULL) {
-                    fclose(input_fp);
-                    input_fp = NULL;
-                }
-                if (loop) {
-                    input_fp = fopen(file_name, "rb");
-                    if (input_fp == NULL) {
-                        printf("Failed to open input file for infinite loop: %s\n",
-                               file_name);
-                        break;
-                    }
-                    if (read_test_data(input_fp, buf, frame_size) < 0) {
-                        break;
-                    }
-                } else {
-                    break;
-                }
-            }
+        clock_gettime(CLOCK_REALTIME, &ts_recv);
+        if (first_frame) {
+            ts_begin = ts_recv;
+            first_frame = false;
         }
 
-        err = mesh_put_buffer(&buf);
-        if (err) {
-            printf("Failed to put buffer: %s (%d)\n", mesh_err2str(err), err);
-            break;
+        if (dump_fp) {
+            fwrite(buf->data, buf->data_len, 1, dump_fp);
+        } else {
+            /* Following code are mainly for test purpose, it requires the sender side to
+               pre-set the first several bytes */
+            ptr = buf->data;
+            if (*(uint32_t *)ptr != frame_count) {
+                printf("Wrong data content: expected %u, got %u\n", frame_count, *(uint32_t*)ptr);
+                /* catch up the sender frame count */
+                frame_count = *(uint32_t*)ptr;
+            }
+            ptr += sizeof(frame_count);
+            ts_send = *(struct timespec *)ptr;
+
+            latency = 1000.0 * (ts_recv.tv_sec - ts_send.tv_sec);
+            latency += (ts_recv.tv_nsec - ts_send.tv_nsec) / 1000000.0;
         }
 
         if (frame_count % stat_interval == 0) {
@@ -362,30 +322,24 @@ int main(int argc, char** argv)
 
             clock_gettime(CLOCK_REALTIME, &ts_begin);
         }
-
-        printf("TX frames: [%d], FPS: %0.2f [%0.2f]\n", frame_count, fps, vid_fps);
+        printf("RX frames: [%u], latency: %0.1f ms, FPS: %0.3f\n", frame_count, latency, fps);
         printf("Throughput: %.2lf MB/s, %.2lf Gb/s \n", throughput_MB,  throughput_MB * 8 / 1000);
 
         frame_count++;
 
-        if (frame_count >= total_num) {
+        err = mesh_put_buffer(&buf);
+        if (err) {
+            printf("Failed to put buffer: %s (%d)\n", mesh_err2str(err), err);
             break;
         }
-
-        /* Timestamp for frame end. */
-        clock_gettime(CLOCK_REALTIME, &ts_frame_end);
-
-        /* sleep for 1/fps */
-        __useconds_t spend = 1000000 * (ts_frame_end.tv_sec - ts_frame_begin.tv_sec) + (ts_frame_end.tv_nsec - ts_frame_begin.tv_nsec)/1000;
-        printf("pacing: %d\n", pacing);
-        printf("spend: %d\n", spend);
 
         printf("\n");
     }
 
-    sleep(2);
-
     /* Clean up */
+    if (dump_fp)
+        fclose(dump_fp);
+
     err = mesh_delete_connection(&conn);
     if (err)
         printf("Failed to delete connection: %s (%d)\n", mesh_err2str(err), err);
@@ -393,10 +347,6 @@ int main(int argc, char** argv)
     err = mesh_delete_client(&client);
     if (err)
         printf("Failed to delete mesh client: %s (%d)\n", mesh_err2str(err), err);
-
-    if (input_fp != NULL) {
-        fclose(input_fp);
-    }
 
     return 0;
